requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.1.0=gpu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py39h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=3.5.0=py39h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
astor=0.8.1=py39h06a4308_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py39h06a4308_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py39h06a4308_0
brotlipy=0.7.0=py39h27cfd23_1003
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.9.24=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
cryptography=37.0.1=py39h9ce1e76_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dlib=19.24.0=pypi_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
face-recognition=1.3.0=pypi_0
face-recognition-models=0.3.0=pypi_0
fftw=3.3.9=h27cfd23_1
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py39h7f8727e_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py39hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py39hec9cf62_0
hdf5=1.10.6=h3ffc7dd_1
icu=58.2=he6710b0_3
idna=3.4=py39h06a4308_0
importlib-metadata=4.11.3=py39h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py39h06a4308_0
ipython=8.4.0=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.16.0=py39h06a4308_0
jupyter=1.0.0=py39h06a4308_8
jupyter_client=7.3.5=py39h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.11.1=py39h06a4308_0
jupyter_server=1.18.1=py39h06a4308_0
jupyterlab=3.4.4=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.15.2=py39h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
krb5=1.19.2=hac12032_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
loguru=0.6.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py39h06a4308_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-inline=0.1.6=py39h06a4308_0
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=py39h5eee18b_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py39h06a4308_0
nbconvert=6.4.4=py39h06a4308_0
nbformat=5.5.0=py39h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py39h06a4308_0
notebook=6.4.12=py39h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.23.3=py39h14f4228_0
numpy-base=1.23.3=py39h31eccc5_0
oauthlib=3.2.1=py39h06a4308_0
openssl=1.1.1q=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=pypi_0
pip=22.2.2=py39h06a4308_0
ply=3.11=py39h06a4308_0
prometheus_client=0.14.1=py39h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.20.1=py39h295c915_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py39h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.7=py39h6a678d5_1
pyqt5-sip=12.11.0=py39h6a678d5_1
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.13=haa1d7c7_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39h06a4308_0
python-flatbuffers=2.0=pyhd3eb1b0_0
pytz=2022.1=py39h06a4308_0
pyzmq=23.2.0=py39h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtconsole=5.3.2=py39h06a4308_0
qtpy=2.2.0=py39h06a4308_0
qtwebkit=5.212=h4eab89a_4
readline=8.1.2=h7f8727e_1
requests=2.28.1=py39h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.9.1=py39h14f4228_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py39h06a4308_0
sip=6.6.2=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py39h06a4308_1
soupsieve=2.3.2.post1=py39h06a4308_0
sqlite=3.39.3=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.9.0=py39h06a4308_0
tensorboard-data-server=0.6.0=py39hca6d32c_0
tensorboard-plugin-wit=1.8.1=py39h06a4308_0
tensorflow=2.4.1=gpu_py39h8236f22_0
tensorflow-base=2.4.1=gpu_py39h29c2da4_0
tensorflow-estimator=2.6.0=pyh7b7c402_0
tensorflow-gpu=2.4.1=h30adc30_0
termcolor=1.1.0=py39h06a4308_1
terminado=0.13.1=py39h06a4308_0
testpath=0.6.0=py39h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.2=py39h5eee18b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py39h06a4308_0
typing_extensions=4.3.0=py39h06a4308_0
tzdata=2022e=h04d1e81_0
urllib3=1.26.12=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
websocket-client=0.58.0=py39h06a4308_4
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39h06a4308_0
wrapt=1.14.1=py39h5eee18b_0
xz=5.2.6=h5eee18b_0
yarl=1.8.1=py39h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py39h06a4308_0
zlib=1.2.12=h5eee18b_3
zstd=1.5.2=ha4553b6_0
